refactor(run): isolate release kit planning
